requirements.txt
python-telegram-bot==20.6
psycopg2==2.9.9
